+++ test_renderer.py
@@ -45,14 +45,14 @@ def test_ok_including_a_Tag_in_statics():
     assert "<mystyle>" in str(r) # will not be included
     # and a logger.warning is outputed
 
-# def test_statics_in_real_statics():
-#     s1=Tag.div( [1,"h",Tag.div("dyn"),Tag.H.div("dyn",_id=12)] )
-#     s2=Tag.div( [1,"h",Tag.H.div("stat"),Tag.H.Section( Tag("yolo") )] )
+def test_statics_in_real_statics():
+    s1=Tag.H.div( [1,"h",Tag.div("dyn"),Tag.H.div("dyn",_id=12)] )
+    s2=Tag.div( [1,"h",Tag.H.div("stat"),Tag.H.Section( Tag("yolo") )] )
 
-#     assert "id=" in str(s1)
-#     assert s1._renderStatic() =='<div>1h<div>dyn</div><div id="12">dyn</div></div>'
-#     assert "id=" in str(s2)
-#     assert s2._renderStatic() =="<div>1h<div>stat</div><Section><div>yolo</div></Section></div>"
+    assert "id=" in str(s1)
+    assert str(s1._ensureTagBase()) =='<div>1h<div>dyn</div><div id="12">dyn</div></div>'
+    assert "id=" in str(s2)
+    assert str(s2._ensureTagBase()) =="<div>1h<div>stat</div><Section><div>yolo</div></Section></div>"
 
 
 def test_render_a_tag_with_interaction():
@@ -493,6 +493,52 @@ def test_discovering_css():
     r=HRenderer(OOOL(),"//js interact")
     test(r)
 
+def test_imports():
+
+    class Ya(Tag.div):
+        statics = Tag.H.style("""body {font-size:100px}""", _id="Ya")
+
+        def __init__(self):
+            super().__init__()
+            self <= "Ya"
+
+
+    class Yo(Tag.div):
+        statics = Tag.H.style("""body {background:#CFC;}""", _id="Yo")
+
+        def __init__(self):
+            super().__init__()
+            self <= "Yo"
+
+
+    class AppWithImport(Tag.body):
+        statics = Tag.H.style("""body {color: #080}""", _id="main")
+        imports = Yo
+
+        def __init__(self):
+            super().__init__()
+
+            self <= Yo()
+
+    class AppWithoutImport(Tag.body):
+        statics = Tag.H.style("""body {color: #080}""", _id="main")
+
+        def __init__(self):
+            super().__init__()
+
+            self <= Yo()
+
+    html=str(HRenderer( AppWithImport(), ""))
+    styles = re.findall("(<style[^<]+</style>)",html)
+    assert len(styles)==2
+
+    html=str(HRenderer( AppWithoutImport(), ""))
+    styles = re.findall("(<style[^<]+</style>)",html)
+    assert len(styles)>2
+
+
+
+
 if __name__=="__main__":
     # test_ko_try_render_a_tagbase()
     # test_render_a_tag_with_interaction()
@@ -510,4 +556,5 @@ if __name__=="__main__":
 
     # test_intelligent_rendering()
     # test_statics_in_real_statics()
-    test_render_yield_with_scripts()
+    # test_render_yield_with_scripts()
+    test_statics_in_real_statics()